Develop

See merge request projetm2/backend!9

+++ src/Union.Backend.API/Controllers/GardenScoresController.cs
@@ -1,8 +1,11 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Union.Backend.Service.Services;
-using System;
+using Union.Backend.Service.Exceptions;
 using Union.Backend.Service.Dtos;
+using Union.Backend.Service.Auth;
+using System.Net;
 
 namespace Union.Backend.API.Controllers
 {
@@ -11,6 +14,7 @@ namespace Union.Backend.API.Controllers
     public class GardenScoresController : ControllerBase
     {
         private readonly ScoresService service;
+        private readonly GardensService GardenService;
         public GardenScoresController(ScoresService service)
         {
             this.service = service;
@@ -23,22 +27,63 @@ namespace Union.Backend.API.Controllers
         }
 
 
-        [HttpPost("{id}/score")] //TODO
+        [HttpPost("{id}/score")]
         public async Task<IActionResult> AddScore([FromRoute(Name = "id")] Guid GardenId, [FromBody] ScoreDto Score)
         {
-            return Created("TODO", await service.AddScore(Score, GardenId));
+            try
+            {
+                var id = Utils.ExtractIdFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]);
+                var garden = await GardenService.GetGardenById(GardenId);
+                if (garden.Data.Owner != id && garden.Data.Owner != id && !Utils.IsAdminRoleFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]))
+                {
+                    return Forbid();
+                }
+                var result = await service.AddScore(Score, GardenId);
+                return Created($"api/gardens/{result.Data.Id}/score", result);
+
+            }
+            catch (HttpResponseException)
+            {
+                throw;
+            }
+            catch (Exception)
+            {
+                throw new BadRequestApiException();
+            }
+
+
         }
 
-        [HttpPost("score/{id}/report")] //TODO
+        [HttpPost("score/{id}/report")]
         public async Task<IActionResult> ReportScore([FromRoute(Name = "id")] Guid ScoreId, [FromBody] ScoreDto Score)
         {
-            return Created("TODO", await service.ReportScore(ScoreId, Score));
+            return Ok(await service.ReportScore(ScoreId, Score));
         }
 
         [HttpDelete("score/{id}")]
         public async Task DeleteScore([FromRoute(Name = "id")] Guid ScoreId)
         {
-            await service.DeleteScore(ScoreId);
+
+            try
+            {
+                var id = Utils.ExtractIdFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]);
+                var score = await service.GetScore(ScoreId);
+                if (score.Data.Rater == id || Utils.IsAdminRoleFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]))
+                {
+                    await service.DeleteScore(ScoreId);
+                }
+
+            }
+            catch (HttpResponseException)
+            {
+                throw;
+            }
+            catch (Exception)
+            {
+                throw new BadRequestApiException();
+            }
+
+
         }
 
     }
