+++ src/Union.Backend.API/Controllers/UserScoresController.cs
@@ -1,12 +1,15 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Union.Backend.Service.Services;
-using System;
+using Union.Backend.Service.Exceptions;
 using Union.Backend.Service.Dtos;
+using Union.Backend.Service.Auth;
+using System.Net;
 
 namespace Union.Backend.API.Controllers
 {
-    [Route("api/users/")]
+    [Route("api/Users")]
     [ApiController]
     public class UserScoresController : ControllerBase
     {
@@ -23,10 +26,11 @@ namespace Union.Backend.API.Controllers
         }
 
 
-        [HttpPost("{id}/score")] //TODO
+        [HttpPost("{id}/score")]
         public async Task<IActionResult> AddScore([FromRoute(Name = "id")] Guid UserId, [FromBody] ScoreDto Score)
         {
-            return Created("TODO", await service.AddScore(Score, UserId));
+            var result = await service.AddScore(Score, UserId);
+            return Created($"/api/Users/{result.Data.Id}/score", result);
         }
 
         [HttpPost("score/{id}/report")]
@@ -38,7 +42,31 @@ namespace Union.Backend.API.Controllers
         [HttpDelete("score/{id}")]
         public async Task DeleteScore([FromRoute(Name = "id")] Guid ScoreId)
         {
-            await service.DeleteScore(ScoreId);
+            try
+            {
+                var id = Utils.ExtractIdFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]);
+                var score = await service.GetScore(ScoreId);
+
+                if (score.Data.Rater == id || Utils.IsAdminRoleFromToken(Request.Headers[HttpRequestHeader.Authorization.ToString()]))
+                {
+                    await service.DeleteScore(ScoreId);
+                }
+                else
+                {
+                    throw new ForbidenException();
+                }
+
+            }
+            catch (HttpResponseException)
+            {
+                throw;
+            }
+            catch (Exception)
+            {
+                throw new BadRequestApiException();
+            }
+
+
         }
 
     }
